FIX: Under Windows, do not use Inkscape for .svg->.png conversion
There is a bug in Inkscape that makes it crash when trying such
conversions. Under windows the .png thumbnails are generated using
MATLAB's built-in .png exporter. This should be an acceptable solution
as long as MATLAB is not run with -nodisplay, which is a rare
case under Windows. Once the bug in Inkscape has been solved the
.png generation will be again taken care by Inkscape.