MAJOR FIX: import epoched .mff files
Previous versions of the .mff importer ignored completely the possible
presence of epochs within the .mff file. As a result, the imported
events may be placed in the wrong data sample if (1) multiple epochs
exist in the file and (2) there is a time gap between epochs. The
new version should handle correctly the presence of multiple epochs
by appropriately shifting backwards the sample positions of the events.

Additionally, the presence of multiple epochs within an .mff file
is encoded in the imported physioset as follows:

* Property Time of all physioset events contain the absolute times
  of those events, as a MATLAB datenum.

* If two or more epochs are found, a matching number of epoch_begin
  events are added to the physioset. Such events are located in the
  boundaries between epochs and their Time property contains the
  absolute begin time of the epoch.

* You can retrieve the actual sampling time (relative to the beginning
  of the recording) for each physioset data sample using method
  get_sampling_time().